Add a check whether Wake-On-Lan is enabled for Realtek ethernet controllers
Some bugs have been reported with low hardware sleep residency specifically
when WoL is disabled.  This is certainly either a firmware or kernel bug, but
warn users for now that they could be encountering it to give pointers.

This will be needlessly noisy for some users, but as clairvoyance is granted
on the actual root cause of the problem this check can be amended to detect
more of the situation and only show the message as necessary.