Add a check for artificially limiting max cores
If the kernel has been compiled with a lower number of CONFIG_NR_CPUS
than the number of cores actually on the system this will prevent the
kernel from being able to put some cores into the proper power state
for hardware sleep to work.

Use the hardware topology CPUIID to detect this and fail the suspend.

Link: https://www.amd.com/content/dam/amd/en/documents/processor-tech-docs/programmer-references/24593.pdf
Link: https://www.amd.com/content/dam/amd/en/documents/processor-tech-docs/programmer-references/24594.pdf