Capture ACPI Notify() debugging with @pm_debugging decorator
When an APU leaves HW sleep becauce of the EC SCI, the kernel will
process all pending EC events which will then often lead to ACPI Notify().

Although potentially noisy; this information is really useful to
understand what is going on for a given sleep cycle.

When available (kernel is compiled with `CONFIG_ACPI_DEBUG`) capture
this information for each cycle by turning on debug level and debug
layer.

if this turns out to be too noisy we can add a configuration knob
to the script to control whether it's enabled.

Signed-off-by: Mario Limonciello <mario.limonciello@amd.com>