amd_s2idle: Write `wakeup_count` before suspending the system
The purpose of the `wakeup_count` attribute is to avoid a race condition
when suspending the system where an event to wake it comes in while it's
being suspended.  Use this attribute as described in kernel source code
to suspend to avoid triggering -EIO errors that could occur causing an
abort.

Link: https://git.kernel.org/torvalds/c/c125e96f04442
Signed-off-by: Mario Limonciello <mario.limonciello@amd.com>