amd_s2idle: Convert to be part of the python wheel
This adjusts a lot about how amd_s2idle works.  It can be run
on demand to do a test, and results are stored into a sqlite database.

If desirable users can install a systemd hook that will gather debugging
data on every suspend cycle.  Reports can then be generated later on
to compare data from cycles.

Speaking of reports; they can now be done in text, markdown or html.
The default is HTML.

Reports include graphs and javascript that can make the page a lot
nicer to interact with.

Because this is a wheel it only gets installed to a user profile by
default.  This means it's a lot more awkward to run as root, but root
is important.  Add some special handling if not running as root to at
least attempt to use sudo to escalate privs.  This is only needed
when running a test locally instead of with systemd.

Signed-off-by: Mario Limonciello <mario.limonciello@amd.com>